refactor: move osnap to osnap directory
